implemented the concept of rounds of confussionDiffision for the singular 768x768. as suggested by the paper, 5 rounds have been defined and have been run. all the output intermediate frames have also been written and are in this commit. also the decription steps has been performed with the appropriate inverse operations. the final decripted image after the round is found after the final inverse confusion round : afterInvConfusion.png. encryption process breaks if global confusion seed sc value is higher then 1e10 in magnitude

+++ include/prbg_main_plcm.hpp
@@ -4,6 +4,6 @@
 #include <vector>
 #include <iostream>
 
-std::vector<double> generatePRBGMainKeys(double x0, double p, int numParameters4subsequentPRBGas, int *sc);
+std::vector<double> generatePRBGMainKeys(double x0, double p, int numParameters4subsequentPRBGas, uint64_t *sc);
 
 #endif
